made open_vifm async
added a jobstart to run vifm as a async process to make it less laggy

+++ lua/config/customfunctions.lua
@@ -3,21 +3,32 @@ function Foo()
 	print("Test function Runs!")
 end
 
+-------------------------------------------------------------------------------------------------------
+-- Function to open VIFM
 local function open_vifm()
-	local tmpfile = "/tmp/vifm_selected"         -- Temporary file to store the selected file
-	os.execute("vifm --choose-files " .. tmpfile) -- Open vifm and store the selection
+	local tmpfile = "/tmp/vifm_selected"
 
-	local file = vim.fn.readfile(tmpfile)[1]     -- Read the first selected file from the tmpfile
-	if file and file ~= "" then
-		vim.cmd("edit " .. vim.fn.fnameescape(file)) -- Open the file in a new buffer
-	end
+	-- Command to open Terminal and run vifm
+	local terminal_cmd = "$TERM -e vifm --choose-files " .. tmpfile
 
-	os.remove(tmpfile) -- Clean up the temporary file
+	-- Run the command
+	vim.fn.jobstart(terminal_cmd, { --jobstart for async VIFM instance makes it less laggy
+		on_exit = function(_, exit_code)
+			if exit_code == 0 then
+				-- Read the selected file after vifm exits
+				local file = vim.fn.readfile(tmpfile)[1]
+				if file and file ~= "" then
+					vim.cmd("edit " .. vim.fn.fnameescape(file))
+				end
+				os.remove(tmpfile) -- Clean up the temporary file
+			end
+		end,
+	})
 end
-
 vim.api.nvim_create_user_command("VifmOpen", open_vifm, {}) -- Create a :VifmOpen command
 
-
+-------------------------------------------------------------------------------------------------------
+-- Function to open lazygit
 local function open_lazygit()
 	os.execute("footclient -e lazygit &")
 end
